+++ kamehouse-shell/bin/common/backup/backup-server-functions.sh
@@ -135,13 +135,6 @@ backupHomeFolders() {
   pullDockerHomeFolders
 
   if ${IS_LINUX_HOST}; then
-    sudo cp -vrf ${HOME}/home-synced ${DEST_HOME}/
-  else
-    cp -vrf ${HOME}/home-synced ${DEST_HOME}/
-  fi
-  checkCommandStatus "$?" "An error occurred during file copy"
-
-  if ${IS_LINUX_HOST}; then
     sudo cp -vrf ${HOME}/.kamehouse ${DEST_HOME}/
   else
     cp -vrf ${HOME}/.kamehouse ${DEST_HOME}/
@@ -151,9 +144,8 @@ backupHomeFolders() {
 
 pullDockerHomeFolders() {
   log.info "Pulling folders to sync from docker prod container, if it's running"
-  scp -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced ${HOME}/home-synced/docker/ 
-  scp -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse ${HOME}/home-synced/docker/ 
-  rm -f ${HOME}/home-synced/docker/.kamehouse/.kamehouse-docker-container-env
+  scp -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse ${HOME}/.kamehouse/docker/ 
+  rm -f ${HOME}/.kamehouse/docker/.kamehouse-docker-container-env
 }
 
 backupWorkspaceEclipse() {