+++ kamehouse-shell/bin/kamehouse/docker/docker-reinit-container-data-from-host.sh
@@ -35,7 +35,6 @@ mainProcess() {
   else
     reinitSsh
     reinitKameHouseFolder
-    reinitHomeSynced
     reinitMariadb
   fi
 }
@@ -81,54 +80,51 @@ reinitKameHouseFolder() {
     log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/.cred /home/${DOCKER_USERNAME}/.kamehouse/.shell/\""
     ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/.cred /home/${DOCKER_USERNAME}/.kamehouse/.shell/"
 
-    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/.*.pwd.enc /home/${DOCKER_USERNAME}/.kamehouse\""
-    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/.*.pwd.enc /home/${DOCKER_USERNAME}/.kamehouse"
+    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/.*.pwd.enc /home/${DOCKER_USERNAME}/.kamehouse/keys\""
+    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/.*.pwd.enc /home/${DOCKER_USERNAME}/.kamehouse/keys"
   else
     log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/.shell/.cred ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/.shell/"
     scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/.shell/.cred ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/.shell/
 
-    log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/.*.pwd.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse"
-    scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/.*.pwd.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse
+    log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/keys/.*.pwd.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/keys"
+    scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/keys/.*.pwd.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/keys
   fi
-}
 
-reinitHomeSynced() {
-  log.info "Setup home-synced folder"
   if [ "${DATA_SOURCE}" == "docker-init" ]; then
-    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/integration-test-cred.enc /home/${DOCKER_USERNAME}/home-synced/.kamehouse\""
-    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/integration-test-cred.enc /home/${DOCKER_USERNAME}/home-synced/.kamehouse"
+    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/integration-test-cred.enc /home/${DOCKER_USERNAME}/.kamehouse/keys\""
+    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/docker/keys/integration-test-cred.enc /home/${DOCKER_USERNAME}/.kamehouse/keys"
 
-    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.pkcs12 /home/${DOCKER_USERNAME}/home-synced/.kamehouse/keys/kamehouse.pkcs12\""
-    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.pkcs12 /home/${DOCKER_USERNAME}/home-synced/.kamehouse/keys/kamehouse.pkcs12"
+    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.pkcs12 /home/${DOCKER_USERNAME}/.kamehouse/keys/kamehouse.pkcs12\""
+    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.pkcs12 /home/${DOCKER_USERNAME}/.kamehouse/keys/kamehouse.pkcs12"
 
-    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.crt /home/${DOCKER_USERNAME}/home-synced/.kamehouse/keys/kamehouse.crt\""
-    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.crt /home/${DOCKER_USERNAME}/home-synced/.kamehouse/keys/kamehouse.crt"
+    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.crt /home/${DOCKER_USERNAME}/.kamehouse/keys/kamehouse.crt\""
+    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-commons-core/src/test/resources/commons/keys/sample.crt /home/${DOCKER_USERNAME}/.kamehouse/keys/kamehouse.crt"
   else
-    log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/.kamehouse/integration-test-cred.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/.kamehouse"
-    scp -C -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/.kamehouse/integration-test-cred.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/.kamehouse
+    log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/keys/integration-test-cred.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/keys"
+    scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/keys/integration-test-cred.enc ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/keys
 
-    log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/.kamehouse/keys/* ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/.kamehouse/keys"
-    scp -C -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/.kamehouse/keys/* ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/.kamehouse/keys
+    log.debug "scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/keys/* ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/keys"
+    scp -C -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/keys/* ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/keys
   fi
   
   case ${DATA_SOURCE} in
   "none")
-    log.info "Skipping setup of home-synced/mariadb"
+    log.info "Skipping setup of .kamehouse/mariadb"
     ;;
   "docker-init")
     log.info "Resetting mariadb dump data from initial docker container data"
-    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"mkdir -p /home/${DOCKER_USERNAME}/home-synced/mariadb/dump/old ; cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-shell/bin/kamehouse/sql/mariadb/dump-kamehouse.sql /home/${DOCKER_USERNAME}/home-synced/mariadb/dump\""
-    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "mkdir -p /home/${DOCKER_USERNAME}/home-synced/mariadb/dump/old ; cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-shell/bin/kamehouse/sql/mariadb/dump-kamehouse.sql /home/${DOCKER_USERNAME}/home-synced/mariadb/dump"
+    log.debug "ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C \"mkdir -p /home/${DOCKER_USERNAME}/.kamehouse/mariadb/dump/old ; cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-shell/bin/kamehouse/sql/mariadb/dump-kamehouse.sql /home/${DOCKER_USERNAME}/.kamehouse/mariadb/dump\""
+    ssh -p ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost -C "mkdir -p /home/${DOCKER_USERNAME}/.kamehouse/mariadb/dump/old ; cp -v -f /home/${DOCKER_USERNAME}/git/kamehouse/kamehouse-shell/bin/kamehouse/sql/mariadb/dump-kamehouse.sql /home/${DOCKER_USERNAME}/.kamehouse/mariadb/dump"
     ;;
   "docker-backup")
-    log.info "Exporting mariadb data from ${HOME}/home-synced/docker/mariadb to the container"
-    log.debug "scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/docker/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/"
-    scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/docker/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/
+    log.info "Exporting mariadb data from ${HOME}/.kamehouse/docker/mariadb to the container"
+    log.debug "scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/docker/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/"
+    scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/docker/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/
     ;;
   "host-backup")
-    log.info "Exporting mariadb data from ${HOME}/home-synced/mariadb to the container"
-    log.debug "scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/"
-    scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/home-synced/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/
+    log.info "Exporting mariadb data from ${HOME}/.kamehouse/mariadb to the container"
+    log.debug "scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/"
+    scp -C -r -P ${DOCKER_PORT_SSH} ${HOME}/.kamehouse/mariadb ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/
     ;;
   *) ;;
   esac