+++ kamehouse-shell/bin/kamehouse/mariadb-dump-kamehouse-docker.sh
@@ -46,10 +46,10 @@ exportMariadbDataOnDocker() {
 
 copyDataFromContainerToHost() {
 	log.info "Exporting data from container to host"
-  mkdir -p ${HOME}/home-synced/docker/mariadb
-  rm -rf ${HOME}/home-synced/docker/mariadb
-  log.debug "scp -C -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/mariadb ${HOME}/home-synced/docker/mariadb"
-  scp -C -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/home-synced/mariadb ${HOME}/home-synced/docker/mariadb
+  rm -rf ${HOME}/.kamehouse/docker/mariadb
+  mkdir -p ${HOME}/.kamehouse/docker/mariadb
+  log.debug "scp -C -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/mariadb ${HOME}/.kamehouse/docker/mariadb"
+  scp -C -r -P ${DOCKER_PORT_SSH} ${DOCKER_USERNAME}@localhost:/home/${DOCKER_USERNAME}/.kamehouse/mariadb ${HOME}/.kamehouse/docker/mariadb
 }
 
 parseArguments() {