+++ kamehouse-shell/bin/kamehouse/mariadb-dump-kamehouse.sh
@@ -11,7 +11,7 @@ source ${HOME}/.kamehouse/.shell/.cred
 
 # Global variables
 LOG_PROCESS_TO_FILE=true
-PATH_DUMP_FILE=${HOME}/home-synced/mariadb/dump 
+PATH_DUMP_FILE=${HOME}/.kamehouse/mariadb/dump 
 NUMBER_OF_EXPORTS=3
 DUMP_FILENAME=dump-kamehouse.sql
 LOG_FILENAME=dump-kamehouse.log