+++ kamehouse-shell/bin/kamehouse/mariadb-restore-kamehouse.sh
@@ -10,7 +10,7 @@ fi
 source ${HOME}/.kamehouse/.shell/.cred
 
 LOG_PROCESS_TO_FILE=true
-MARIADB_DUMP_FILE=${HOME}/home-synced/mariadb/dump/dump-kamehouse.sql
+MARIADB_DUMP_FILE=${HOME}/.kamehouse/mariadb/dump/dump-kamehouse.sql
 
 mainProcess() {
   if [ -f "${MARIADB_DUMP_FILE}" ]; then